# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Туристический ассистент API",
    "description": "REST API веб-интерфейса: multi-user SaaS, поездки в SQLite, асинхронная сборка программы LangGraph, BYOK OpenRouter.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/trips": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List Trips",
        "operationId": "list_trips_api_trips_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TripSummaryResponse"
                  },
                  "type": "array",
                  "title": "Response List Trips Api Trips Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create Trip",
        "operationId": "create_trip_api_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/trips/{trip_id}": {
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Delete Trip",
        "operationId": "delete_trip_api_trips__trip_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip",
        "operationId": "get_trip_api_trips__trip_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/program": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Program",
        "operationId": "get_program_api_trips__trip_id__program_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/hotel-zones": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Hotel Zones",
        "description": "Зоны поиска отелей Booking.com вдоль выбранного маршрута.",
        "operationId": "get_hotel_zones_api_trips__trip_id__hotel_zones_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          },
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelZonesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/program/feedback": {
      "put": {
        "tags": [
          "trips"
        ],
        "summary": "Set Program Feedback",
        "operationId": "set_program_feedback_api_trips__trip_id__program_feedback_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/affiliate-clicks": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Log Affiliate Click",
        "description": "Локальный учёт клика по affiliate-ссылке в блоке билетов.",
        "operationId": "log_affiliate_click_api_trips__trip_id__affiliate_clicks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AffiliateClickRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/preferences": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_trips__trip_id__preferences_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TripPreferences"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Preferences Api Trips  Trip Id  Preferences Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/runs": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Start Run",
        "operationId": "start_run_api_trips__trip_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/review": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Submit Review",
        "operationId": "submit_review_api_trips__trip_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AffiliateClickRequest": {
        "properties": {
          "target_url": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 8,
            "title": "Target Url"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "target_url"
        ],
        "title": "AffiliateClickRequest"
      },
...
      "CreateTripRequest": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City"
          },
          "dates": {
            "type": "string",
            "title": "Dates"
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City",
            "default": "Москва"
          },
          "user_query": {
            "type": "string",
            "title": "User Query",
            "default": "Составь культурную программу поездки"
          },
          "preferences": {
            "$ref": "#/components/schemas/TripPreferences"
          },
          "start_run": {
            "type": "boolean",
            "title": "Start Run",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "city",
          "dates",
          "preferences"
        ],
        "title": "CreateTripRequest"
      },
      "CreateTripResponse": {
        "properties": {
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "trip_id"
        ],
        "title": "CreateTripResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotelZonesResponse": {
        "properties": {
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "checkin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin"
          },
          "checkout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout"
          },
          "guests_adults": {
            "type": "integer",
            "title": "Guests Adults"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/HotelZoneResponse"
            },
            "type": "array",
            "title": "Zones"
          },
          "widget_configured": {
            "type": "boolean",
            "title": "Widget Configured",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "case_id",
          "city",
          "guests_adults"
        ],
        "title": "HotelZonesResponse"
      },
...
      "ProgramResponse": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "version_id": {
            "type": "integer",
            "title": "Version Id"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "program": {
            "$ref": "#/components/schemas/FinalProgram"
          },
          "sections": {
            "$ref": "#/components/schemas/StructuredProgramResponse"
          }
        },
        "type": "object",
        "required": [
          "version",
          "version_id",
          "scope",
          "approved",
          "program",
          "sections"
        ],
        "title": "ProgramResponse"
      },
...
      "TripDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "dates": {
            "type": "string",
            "title": "Dates"
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City"
          },
          "user_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Query"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city",
          "dates",
          "origin_city",
          "user_query",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TripDetailResponse"
      },
...
      "TripSummaryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "dates": {
            "type": "string",
            "title": "Dates"
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city",
          "dates",
          "origin_city",
          "status",
          "updated_at"
        ],
        "title": "TripSummaryResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}